Commit large session to attempt other changes
